Add lodash library to the project. Make able to delete items from custom To do lists

+++ app.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const bodyParser = require('body-parser');
 const mongoose = require('mongoose');
+const _ = require('lodash');
 
 mongoose.set('strictQuery', true);
 
@@ -76,7 +77,7 @@ app.get('/', (req, res) => {
 
 app.get('/:customListName', (req, res) => {
 
-    const customListName = req.params.customListName;
+    const customListName = _.capitalize(req.params.customListName);
 
     List.findOne({
         name: customListName
@@ -124,15 +125,23 @@ app.post('/', (req, res) => {
 
 app.post('/delete', (req, res) => {
     const checkedItemId = req.body.checkbox;
+    const listName = req.body.listName;
 
-    Item.findByIdAndRemove(checkedItemId, (err) => {
-        if (!err) {
-            console.log("Succesfully deleted checked item");
-            res.redirect("/");
-        }
-    })
+    if (listName === "Today") {
+        Item.findByIdAndRemove(checkedItemId, (err) => {
+            if (!err) {
+                console.log("Succesfully deleted checked item");
+                res.redirect("/");
+            }
+        });
+    } else {
+        List.findOneAndUpdate({name: listName}, { $pull: {items: {_id: checkedItemId}} }, (err, foundList) => {
+            if (!err) {
+                res.redirect(`/${listName}`);
+            }
+        });
+    }
 });
-
 
 // About Page
